Add files via upload
Implement the Fraction(PL=Ulamek) class with private numerator(PL = Licznik) and denominator(PL=Mianownik) integer fields. You should overload the operators *, -, +, *=, =, -=, +=, --, ++, <, >, ==, << (writing to the stream) and implement constructors. The denominator cannot be zero. Fractions are to be written in the form: Numerator/Denominator, e.g.: "3/4", "-79/8". Create several objects in the main() function and demonstrate the operation of overloaded operators of the Ulamek class.